Fixed broken ? implementation. Also changed treatment of source
Before, source was tokenized (by a simple split()), but now it is
treated as a single string. That way, tokenization can be performed
inline with the parsing. In rules, literal strings encapsulated in ""
will have the whitespace after them being consumed, up until the first
non-whitespace character, effectively updating the sourceIndex variable
to the index of the first non-whitespace character after the string
consumed. Literal strings can still contain whitespace to be matched
against, but note that if the last literal string was "" as well, and
this new literal string STARTS with whitespace ("  the"), it will never
match.