Execution now respects ignoring rules while in fail state
This is good both for efficiencies sake (skip rules of arbitrary
complexity simply by virtue of being in a fail state), and for
correctness (most rules are written to have total control of the
state of the env, so an erroneously executed rule could return env
from a fail state to a pass state, even though it should NOT exit
a fail state unless it was entered into the fail state while INSIDE
of the rule that takes it out of the fail state). Basically:
Rules are executed the way they should be; skipped if in a fail state,
executed if in a state where executing the rules actually means
something.

tl;dr: Code is more correct now.