Added another example set. Note that this JSON parser cheats
This was a quick draft of a JSON parser that cheats by just being
permissive about whether a section is followed by a comma. This will
be changed in the future, as the goal was to do something realistic
quickly, rather than to do a rock-solid JSON parsing ruleset. Note
the problem is with the rule definition, not the parser itself.