Added some more characters to the litVar rule
Even more characters are probably necessary, but in trying to use this
ruleset to parse other rulesets, failure was only ever on having
forgotten to match some character or another. Once these awful, massive
OR chains are replaced with character ranges, this will be less of a
problem, and not nearly as ugly. For now, everything is very much so
a work in progress.