Modified the Turing Tzarpit recognizer (pegEx6.peg) for AST gen
This is still a work in progress, and will likely lead to slight
changes in the AST generation algorithm, so that the resultant
AST is a little more natural (mostly with changing parentage a bit).
The thinking now is that maybe a "null" node should exist to gather
children before being handed off to a real parent node, to avoid
having the first of a ( * rule ) collection of nodes erroneously
be set too deep in the AST, where as the rest of the nodes are at
their proper depth. Look to the first 'if' in a multi-if
Turing Tzarpit state block in the AST for an example of this.