Fixed bug caused by sharing state in "recursive" '+' calls
Before, the VarContainer class contained a single static bool for
storing state while executing an expression governed by the
one-or-more operator, '+'. If, within an expression governed
by the '+' operator, an expression governed by the '+' operator
were encountered, then the single static state bool would
essentially get clobbered and become meaningless, leading to
incorrect, likely infinite-loop execution of the parser. By
making it instead a dynamic array of bools, we can have a state
for each successively "recursive" execution of the '+' operator,
with an extra static bool to keep track of continuations within
the same '+' execution.